Rename xunit namespaces
