+++ unit_test/MinimalCover.Xunit.Core/Data/ParsedFdsTestData.cs
@@ -1,6 +1,7 @@
 ﻿using System.Collections.Generic;
+using MinimalCover.Core;
 
-namespace MinimalCover.Core.Data.Xunit
+namespace MinimalCover.Xunit.Core.Data
 {
   public class ParsedFdsTestData
   {
