+++ unit_test/MinimalCover.Xunit.Core/Parsers/YamlParserTests.cs
@@ -1,9 +1,11 @@
 ﻿using System;
 using System.Collections.Generic;
 using Xunit;
-using MinimalCover.Core.Data.Xunit;
+using MinimalCover.Core;
+using MinimalCover.Core.Parsers;
+using MinimalCover.Xunit.Core.Data;
 
-namespace MinimalCover.Core.Parsers.Xunit
+namespace MinimalCover.Xunit.Core.Parsers
 {
   public class YamlParserTests
   {
